Initial work for home page main section
