+++ src/Pages/Home/Main/Content.mdx
@@ -2,8 +2,8 @@ import { Typography, Spacing, Button } from 'src/UI-Kits';
 
 <div className="main-section">
   <h1>
-    <Typography variant="title" text="web & ui" component="b" />
-    <Typography variant="title" text="developer" component="span" />
+    <Typography variant="title" text="Amin Pakseresht" component="b" />
+    <Typography variant="title" text="web & ui developer" component="span" />
   </h1>
 </div>
 <Spacing margin={['md', 'auto', 'auto']}>
