+++ src/Pages/Home/Main/MainSection.tsx
@@ -1,10 +1,10 @@
-import React, { FC, useContext, lazy } from 'react';
 import { importMDX } from 'mdx.macro';
+import React, { FC, lazy, useContext } from 'react';
+import { darken, ThemeContext } from 'src/Theme';
+import { Grid, Section } from 'src/UI-Kits';
 
-import { ThemeContext, darken } from 'src/Theme';
-import { Section, Grid } from 'src/UI-Kits';
+import { UIDemos } from './Demos';
 import { StyledDevelopment } from './MainSection.styled';
-import { UIDemos } from './Mobile';
 
 const Content = lazy(() => importMDX('./Content.mdx'));
 
@@ -16,7 +16,7 @@ export const HomeMainSection: FC = () => {
   return (
     <Section variant="left-round" backgroundTheme={[darken(secondary, 0.08), secondary]}>
       <Grid>
-        <Grid.Col>
+        <Grid.Col size={8}>
           <StyledDevelopment>
             <Content />
           </StyledDevelopment>
